% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number representing a state in the dataset}

\item{year}{A number representing the year in the dataset}
}
\value{
This function returns a plot of the number of accidents
for based on the \code{state.num} with \code{year} inputs.
Returns an error message if the state or year do not exist in the dataset.
}
\description{
This function plots the number of accidents in an input year on an input US state map.
This function returns an error message if either the \code{state.num} or
\code{year} do not exist in the dataset
}
\examples{
\dontrun{
fars_map_state('10',2013)
}
}
