% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Loads a .csv file in the dataset}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to the .csv file (string)}
}
\value{
The function returns a data frame from the read .csv file.
If the file does not exist, there is an error message.
}
\description{
The function reads a .csv file named by \code{filename} and returns a data frame.
If the file does not exist, there is an error message.
}
\examples{
\dontrun{
accident_2013 <- fars_read("./data/accident_2013.csv.bz2")
}

}
