% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of numbers representing years need to be read}
}
\value{
This function returns a data frame about the number of accidents
with months in rows, years in columns. Returns a invalid year warning if
every input year that does not exist in the datasets.
}
\description{
Reads files corresponding to the vector of years input into data frames
There is an error message if the input invalid.
}
\examples{
\dontrun{
fars_read_years(2013:2015)
}
}
