% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of integers for years}
}
\value{
This function returns a data frame abot the number of accidents
with months in rows, years in columns.
Returns a warning if input years that does not exist in the datasets.
}
\description{
The function summarizes the number of accidents by months based on input \code{years}.
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
}
}
