% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer represents for a year of a data set.}
}
\value{
Returns a string in a format "accident_<year>.csv.bz2" as a file name
If the input is not integer, there is an error message.
}
\description{
The function creates a filename for a .csv.bz2 file based on the \code{year}
}
\examples{
\dontrun{
makefilename(2013)
}
}
